log: add formatting for log level
This commit adds color and bold text to the log levels, to make them
stand out from the rest of the printout.

As the printf command has some trouble with the escape sequences of the
color and bold formatting, custom logic has been implemented in order to
print fixed width strings with formatting. See

  https://stackoverflow.com/questions/57715240/how-to-format-columns-with-printf-in-bash-when-tput-is-used

for details.