fix(utils): remove debug messages from session names
Debug messages were kept in the functions responsible for generating
session names, resulting in very strange session names being generated
when creating new tmux sessions. This commit removes the debug messages,
and changes the corresponding unit tests to use regular expressions in
favor of partial matching for session names, ensuring that this issue
does not happen again in the future.