fix(workflows): fix local shellcheck execution
When executing "act" locally, the GitHub workflow fails if the bats
submodule has been checked out. This is because the shellcheck execution
includes the bats submodule in the sources to check, and apparently bats
does not pass the shellcheck. Ignore the bats submodule when executing
shellcheck so that the workflow passes also when executing locally.