fix: log to stderr where pipes are used
log.sh recognises whether the current output is a tty, a pipe, or a
file. In certain places in jorp.sh, the log output is recognized as a
pipe, resulting in log statements being nested or included in the main
fzf window, which is highly unwanted.

Explicitly redirect such log statements to stderr.